requirements-test.txt
pytest>=7.4.0
pytest-console-scripts>=1.3.1
pytest-cov>=4.1.0
selenium>=4.16.0
django-allauth>=0.52.0
mock>=5.1.0
beautifulsoup4>=4.12.0
requests>=2.25.0